feat: add rule to run tool in BUILD_WORKSPACE_DIRECTORY (#60)
This PR adds a rule which is identical to `cwd`, except that it runs the
tool in `BUILD_WORKSPACE_DIRECTORY` instead of
`BUILD_WORKING_DIRECTORY`. There are some tools which should always be
run from the module/workspace root and I think it would be useful to add
a convenience rule for this.