Support multiple multitool hubs (#65)
We've supported multiple hubs in WORKSPACE for a while, but bzlmod users were forced to constrain to a single common hub.

This PR maintains existing behavior but treats the common `multitool` hub as the default hub, still registers its toolchains, and now offers the ability to register additional, alternate hubs by name under the same combination rules. It also updates the readme to share how to do this in bzlmod, and calls out that users must invoke `register_toolchains` themselves in order to make the alternate hubs' tools available to Bazel.